add comments + reorganize
